Move cache setup to SteamApiClient; add hybrid cache
Relocates cache provider configuration from RandomSteamGame to the SteamApiClient library and introduces a hybrid cache (L1 in-memory + L2 distributed). Changes:
- Removed NeoSmart.Sqlite and SqlServer cache registration and package refs from RandomSteamGame.
- Added Sqlite/SqlServer packages and NeoSmart.Sqlite using to SteamApiClient csproj and DI.
- SteamApiDependencyInjection now configures SQLite (creates Data folder, uses Data Source filename) or SqlServer distributed cache and registers AddHybridCache with 24h L2 / 30m L1 defaults.
- Keeps Steam options validation and throws on unsupported provider.
Files: ServiceExtensions.cs, RandomSteamGame.csproj, SteamApiClient.csproj, SteamApiDependencyInjection.cs
